service/requirements.txt
django
psycopg2
djangorestframework
django-extensions
django-cors-headers
Pillow